Code clean-ups.
As we are starting to complexify the code, it's time to clean-up the code ...

Discovered a bug where each command was doing a network round-trip to get the server version. Fixed by doing the round-trip for each remote command (ie. the one ine server.ml/RemoteServer).

Tests still pass, but the new things are covered yet, so I've probably break some. Will fix that next week :-)