[install/upgrade] fix the restoration of old configuration Suppose that "o.3" is installed and it exists a binary named "o" in $opam/bin . When the solver proposes to change from "o.3" to "o.2", the binary "o" was wrongly erased and not restored.
Note that "o.2" could also install a binary "o". Here, the difficulty is to restore back the correct binary "o" in $opam/bin that were installed before the [proceed_tochange] execution. Because [proceed_tochange] performs some side effects modifications on variable with [File.Env], one solution is to call it again with a recursive call.