Export OCAML_TOPLEVEL_PATH (=~/.opam/<alias>/lib/toplevel)
The goal is to be able to use 'ocaml -I $OCAML_TOPLEVEL_PATH'

This is not ideal, but better than nothing. I'm a bit relunctant for now on to install a wrapper, around ocaml to automatically install the -I ..., as it may interfere with existing aliases (as people use either 'alias ocaml=ledit ocaml' or 'alias ocaml=rlwrap'). This can be changed later, when I'm convince I will not break anything or when we choose to install an improved toplevel by default.