Simplification of backends code
Now everything is in one file for each backend, with the same BACKEND signature to satisfy - this means it's easier to deploy (as you don't have dozains of binaries to copy at the right place), it's easier to reason about (as you have type constraints, instead of having to read on stdout after calling a script) and it's easier to evolve, as you can easily add new functions if needed.

Currenlty, only the curl and the rsync backend are ported over the new interface, and rsync is passing the test-suite.