[install] new heuristic for the solver installer.
Previously, when trying to install an old package which would trigger a downgrade, the solver heuristic was a bit lost and tried to randomly change some other package versions.

Now the heuristic is the following:
* we first run the solver to find a consistent solution
* then, for all the new packages appearing in the solution:
  - if a solution with the most recent version of the package is consistent, then put the package in a bag
  - if the solution where all the packages in the bag are installed is consistent, then use it
  - otherwise, keep the original solution

This ensure that no independant package is changed when downgrading a subset of packages to fillful the requirement to install a new package.