More clever handling of optional dependencies.
Now the optional dependencies are automatically added to the depends field when it is installed. This ensure that the solver checks that the right package versions are installed, and also it should solve the case where we have an optional conjunction.

For instance:
depopts: [ "async" ("lwt" & "ssh") ]

is translated to:
depends += [ "lwt" & "ssh" ]

if either any version of lwt or ssh is installed.